enforce algo as strategy param
