+++ tests/strategies/test_single_organization.py
@@ -1,28 +1,7 @@
-from contextlib import nullcontext as does_not_raise
-
 import pytest
 
-from substrafl import exceptions
 from substrafl.nodes.train_data_node import TrainDataNode
-from substrafl.schemas import StrategyName
 from substrafl.strategies import SingleOrganization
-
-
-@pytest.mark.parametrize(
-    "strategy_name, expectation",
-    [
-        ("not_the_dummy_strategy", pytest.raises(exceptions.IncompatibleAlgoStrategyError)),
-        (StrategyName.SINGLE_ORGANIZATION, does_not_raise()),
-    ],
-)
-def test_match_algo_single_organization(strategy_name, dummy_algo_class, expectation):
-    class MyAlgo(dummy_algo_class):
-        @property
-        def strategies(self):
-            return [strategy_name]
-
-    with expectation:
-        SingleOrganization(algo=MyAlgo())
 
 
 @pytest.mark.parametrize("additional_orgs_permissions", [set(), {"TestId"}, {"TestId1", "TestId2"}])
